chore: extract observability data types into pkg/observability (Phase 2 prep)
Carve the cross-cutting request-log, stats, and tracing primitives out of
pkg/gateway/ into a new pkg/observability package. Leave Go type aliases in
pkg/gateway so the 77 external importers — including ~50 service tests
asserting on gateway.RequestEntry field shapes — keep compiling unchanged.

Moved:
- pkg/observability/request_log.go: RequestEntry, RequestLog, RequestFilter,
  RequestStats, RequestBroadcaster, plus a new NextRequestID() helper that
  replaces the package-level requestIDCounter the middleware was
  incrementing inline.
- pkg/observability/trace.go: TraceContext, TraceSummary, TimelineSpan,
  TraceStore + helpers, GenerateTraceID/GenerateSpanID. Verbatim from
  pkg/gateway/tracing.go.
- pkg/observability/traceid/: moved from pkg/gateway/traceid/ (sole
  importer was the trace file above; no external users).

Stays in pkg/gateway:
- LoggingMiddleware + LoggingMiddlewareWithOpts and the request parsers
  (detectServiceFromRequest, extractCallerID, serviceFromAuth, etc.).
- responseRecorder, body-capture const, mem/goroutine sampling.
- SetServicePrefixes from PR #36.

pkg/gateway/tracing.go becomes ~25 lines of aliases. pkg/gateway/logging.go
loses ~200 lines of moved data-type definitions but keeps all middleware.

Zero behavior change — aliases preserve type identity, JSON tags, and
method sets. Verified all packages green except the same pre-existing
pkg/platform/store testcontainers/Docker failure that's been failing on
every PR in this branch.

Unblocks Phase 2 (extract reverse proxy from pkg/gateway/ into pkg/edge):
the proxy can now import pkg/observability directly without depending on
pkg/gateway.

Closes #43
